+++ Datamigratie.Server/Features/Map/ZaaktypeMapping/MapZaaktypeDetails/PublicatieNiveauMapping/ShowPublicatieNiveauMappings/Models/PublicatieNiveauMappingResponse.cs
@@ -1,0 +1,7 @@
+namespace Datamigratie.Server.Features.Map.ZaaktypeMapping.MapZaaktypeDetails.PublicatieNiveauMapping.ShowPublicatieNiveauMappings.Models;
+
+public class PublicatieNiveauMappingResponse
+{
+    public required string DetPublicatieNiveau { get; set; }
+    public required string OzVertrouwelijkheidaanduiding { get; set; }
+}