+++ docs/changelog/changelog.md
@@ -8,6 +8,7 @@
 - [Allow full re-run of a migration](https://dimpact.atlassian.net/browse/DATA-255)
 - [Enable performance testing using large data volumes](https://dimpact.atlassian.net/browse/DATA-253)
 - [Allow partial re-run of migration](https://dimpact.atlassian.net/browse/DATA-297)
+- [(QoL) Auto fill all zaaktype mappings with UI button](https://dimpact.atlassian.net/browse/DATA-407)
 
 ## v0.1.0